Reorganise member_facing.slim, cookie_consent
