+++ app/views/shared/_browser_polyfills.slim
@@ -1,0 +1,1 @@
+= javascript_include_tag "//cdn.polyfill.io/v2/polyfill.min.js?features=Intl.~locale.en,Intl.~locale.de,Intl.~locale.fr&rum=1&unknown=polyfill"
