+++ app/views/shared/_champaign_object.slim
@@ -1,0 +1,5 @@
+javascript:
+  window.champaign = {
+    page: #{page_object(@page).to_json.html_safe},
+    countries: #{countries.to_json.html_safe}
+  };
