--- app/views/shared/_countries.slim
@@ -1,2 +1,0 @@
-javascript:
-  window.champaign.countries = #{countries.to_json.html_safe}
