--- app/views/shared/_page_object.slim
@@ -1,3 +1,0 @@
-javascript:
-  window.champaign = window.champaign || {};
-  window.champaign.page = #{page_object(@page).to_json.html_safe}
