+++ app/views/shared/_shareprogress_snippet.slim
@@ -1,0 +1,1 @@
+script type="text/javascript" charset="utf-8" async=true src="//c.shpg.org/99/sp.js"
